fix(api): Allow arbitrary variables on the payload namespace
Allow arbitrary variables on the payload namespace after adding at least one variable in a previous control value.
