+++ apps/api/src/app/workflows-v2/util/jsonToSchema.ts
@@ -1,5 +1,7 @@
 import { JSONSchemaDefinition, JSONSchemaDto } from '@novu/shared';
 
+export { JSONSchemaDto } from '@novu/shared';
+
 export function emptyJsonSchema(): JSONSchemaDto {
   return {
     type: 'object',
@@ -8,70 +10,7 @@ export function emptyJsonSchema(): JSONSchemaDto {
   };
 }
 
-export function convertJsonToSchemaWithDefaults(unknownObject?: Record<string, unknown>) {
-  if (!unknownObject) {
-    return {};
-  }
-
-  return generateJsonSchema(unknownObject) as unknown as JSONSchemaDto;
-}
-
-function generateJsonSchema(jsonObject: Record<string, unknown>): JSONSchemaDto {
-  const schema: JSONSchemaDto = {
-    type: 'object',
-    properties: {},
-    required: [],
-  };
-
-  for (const [key, value] of Object.entries(jsonObject)) {
-    if (schema.properties && schema.required) {
-      schema.properties[key] = determineSchemaType(value);
-      schema.required.push(key);
-    }
-  }
-
-  return schema;
-}
-
-function determineSchemaType(value: unknown): JSONSchemaDto {
-  if (value === null) {
-    return { type: 'null' };
-  }
-
-  if (Array.isArray(value)) {
-    return {
-      type: 'array',
-      items: value.length > 0 ? determineSchemaType(value[0]) : { type: 'null' },
-    };
-  }
-
-  switch (typeof value) {
-    case 'string':
-      return { type: 'string', default: value };
-    case 'number':
-      return { type: 'number', default: value };
-    case 'boolean':
-      return { type: 'boolean', default: value };
-    case 'object':
-      return {
-        type: 'object',
-        properties: Object.entries(value).reduce(
-          (acc, [key, val]) => {
-            acc[key] = determineSchemaType(val);
-
-            return acc;
-          },
-          {} as { [key: string]: JSONSchemaDto }
-        ),
-        required: Object.keys(value),
-      };
-
-    default:
-      return { type: 'null' };
-  }
-}
-
-function isMatchingJsonSchema(schema: JSONSchemaDefinition, obj?: Record<string, unknown> | null): boolean {
+export function isMatchingJsonSchema(schema: JSONSchemaDefinition, obj?: Record<string, unknown> | null): boolean {
   // Ensure the schema is an object with properties
   if (!obj || !schema || typeof schema !== 'object' || schema.type !== 'object' || !schema.properties) {
     return false; // If schema is not structured or no properties are defined, assume match
@@ -101,7 +40,7 @@ function isMatchingJsonSchema(schema: JSONSchemaDefinition, obj?: Record<string,
   return true;
 }
 
-function extractMinValuesFromSchema(schema: JSONSchemaDefinition): Record<string, number> {
+export function extractMinValuesFromSchema(schema: JSONSchemaDefinition): Record<string, number> {
   const result = {};
 
   if (typeof schema === 'object' && schema.type === 'object') {
@@ -121,5 +60,3 @@ function extractMinValuesFromSchema(schema: JSONSchemaDefinition): Record<string
 
   return result;
 }
-
-export { generateJsonSchema, isMatchingJsonSchema, extractMinValuesFromSchema, JSONSchemaDto };
